Prueba 3

+++ main.go
@@ -3,9 +3,9 @@ package main
 import "fmt"
 
 func main() {
-	var base, altura float32
-	fmt.Scanf("%f", &base)
-	fmt.Scanf("%f", &altura)
+	var base, altura int
+	fmt.Scanf("%d", &base)
+	fmt.Scanf("%d", &altura)
 	area := (base*altura)/2
 	fmt.Println(area)
 }
